backend/requirements.txt
# Core Django
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
django-filter==25.1
django-cors-headers==4.7.0

# Database
psycopg2-binary==2.9.10

# Cache
django-redis==5.4.0
redis==5.2.1

# Logging
python-json-logger==2.0.7

# Environment
python-decouple==3.8

# Static files
whitenoise==6.5.0

# Development tools (opcional)
django-extensions==3.2.3
django-debug-toolbar==4.2.0

# Monitoring (opcional)
sentry-sdk[django]==1.32.0

# ASGI Server
uvicorn[standard]==0.32.1
gunicorn==23.0.0

# Testing (dev)
pytest==8.3.4
pytest-django==4.9.0
model-bakery==1.20.0
factory-boy==3.3.1
freezegun==1.5.1

# Utilities
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
Markdown==3.8.2
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
